requirements.txt
mkdocs-material
mkdocs-awesome-pages-plugin
mkdocs-mermaid2-plugin
jinja2
jsonschema
pyyaml